refactor(research): extract ProgressEmitterMixin to eliminate _emit_event duplication
GapDetectionService and DepthExplorationService had identical _emit_event
implementations. Extracted to a shared ProgressEmitterMixin in the shared
package. Also fixes depth_stats["chunks_added"] to include import_chunks_added,
matching the semantics of gap_stats["chunks_added"] and the completion log message.
